Refactor routes helpers
